+++ project/Build.scala
@@ -17,11 +17,11 @@ object TormentaBuild extends Build {
       logLevel in Test := Level.Info
     ) else Seq.empty[Project.Setting[_]]
 
-
   val sharedSettings = extraSettings ++ ciSettings ++ Seq(
     organization := "com.twitter",
     version := "0.5.1",
-    crossScalaVersions := Seq("2.9.2", "2.10.0"),
+    scalaVersion := "2.9.3",
+    crossScalaVersions := Seq("2.9.3", "2.10.0"),
     javacOptions ++= Seq("-source", "1.6", "-target", "1.6"),
     javacOptions in doc := Seq("-source", "1.6"),
     libraryDependencies ++= Seq(
@@ -105,6 +105,8 @@ object TormentaBuild extends Build {
     publishLocal := { }
   ).aggregate(
     tormentaCore,
+    tormentaKestrel,
+    tormentaKafka,
     tormentaTwitter
   )
 
@@ -117,15 +119,18 @@ object TormentaBuild extends Build {
   }
 
   lazy val tormentaCore = module("core").settings(
-    libraryDependencies ++= Seq(
-      "storm" % "storm" % "0.9.0-wip9",
-      "storm" % "storm-kafka" % "0.9.0-wip6-scala292-multischeme",
-      "storm" % "storm-kestrel" % "0.9.0-wip5-multischeme",
-      "com.twitter" %% "chill" % "0.2.1"
-    )
+    libraryDependencies += "storm" % "storm" % "0.9.0-wip9"
   )
 
   lazy val tormentaTwitter = module("twitter").settings(
     libraryDependencies += "org.twitter4j" % "twitter4j-stream" % "3.0.3"
+  ).dependsOn(tormentaCore % "test->test;compile->compile")
+
+  lazy val tormentaKafka = module("kafka").settings(
+    libraryDependencies += "storm" % "storm-kafka" % "0.9.0-wip6-scala292-multischeme"
+  ).dependsOn(tormentaCore % "test->test;compile->compile")
+
+  lazy val tormentaKestrel = module("kestrel").settings(
+    libraryDependencies += "storm" % "storm-kestrel" % "0.9.0-wip5-multischeme"
   ).dependsOn(tormentaCore % "test->test;compile->compile")
 }