+++ tormenta-core/src/main/scala/com/twitter/tormenta/spout/Spout.scala
@@ -17,7 +17,6 @@
 package com.twitter.tormenta.spout
 
 import backtype.storm.topology.IRichSpout
-import com.twitter.bijection.{ AbstractBijection, Bijection, ImplicitBijection }
 
 import java.io.Serializable
 
@@ -29,13 +28,6 @@ import java.io.Serializable
   */
 
 object Spout {
-  implicit def bijection[T, U](implicit bij: ImplicitBijection[T, U])
-      : Bijection[Spout[T], Spout[U]] =
-    new AbstractBijection[Spout[T], Spout[U]] {
-      def apply(spout: Spout[T]) = spout.map(bij(_))
-      override def invert(spout: Spout[U]) = spout.map(bij.invert(_))
-    }
-
   // TODO: Should this be a TravOnce[TravOnce[T]] to test multi-emit?
   def fromTraversable[T](items: TraversableOnce[T]): Spout[T] =
     new BaseSpout[T] {